Param type fixed in change member method

+++ src/java/com/threelittlepigs/codecool/libraryManager/Services/FineService/FineService.java
@@ -18,7 +18,7 @@ public interface FineService {
 
     public void changeStatus(Fine fine, boolean status);
 
-    public void changeMember(Fine fine, int id);
+    public void changeMember(Fine fine, long id);
 
     public List<Fine> getFinesByUserId(int id);
 }
